Rename loggingexporter package for consistency
