+++ auto-exporters/logging/src/main/java/io/opentelemetry/auto/exporters/logging/LoggingExporter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.opentelemetry.auto.exporters.loggingexporter;
+package io.opentelemetry.auto.exporters.logging;
 
 import io.opentelemetry.common.AttributeValue;
 import io.opentelemetry.sdk.trace.data.SpanData;
