+++ auto-exporters/logging/src/main/resources/META-INF/services/io.opentelemetry.sdk.contrib.auto.config.SpanExporterFactory
@@ -1,1 +1,1 @@
-io.opentelemetry.auto.exporters.loggingexporter.LoggingExporterFactory
+io.opentelemetry.auto.exporters.logging.LoggingExporterFactory
